requirements.txt
Flask

# Flask Extensions
Flask-Assets
Flask-DebugToolbar
Flask-SQLAlchemy
Flask-Script
Flask-WTF

# Other
itsdangerous
cssmin
jsmin
gitpython
